WIP: Add verification info to the change screen
Retreive verification info and visualize it on the change screen
as a drop down panel.

It's only rendered as a drop down now but later it can be additionally
rendered aternatively as a panel on change screen directly, like the
custom Javascript solution on review.o.o. And with a setting in user
preferences, exposed by the plugin, user can switch between different
rendering ways. There is also a option to switch the visualization
of the verify informtaion off in this plugin and let custom plugin
to use the data and visualize it in different ways.

Change-Id: Id4980a7a5cf52271c5addfae25a33d49db958043