Add drop down verification info panel to the change screen
Retrieve verification info and visualize it on the change screen
as a drop down panel in change screen header.

It's only rendered as a drop down now but later it can be additionally
rendered aternatively as a panel on change screen directly, like the
custom Javascript solution on review.openstack.org. And with a setting
in user preferences, exposed by the plugin, user can switch between
different rendering ways. There is also an option to switch the
visualization of the verify information off in this plugin and let
custom plugin use the data and visualize it in different ways.

Change-Id: Id4980a7a5cf52271c5addfae25a33d49db958043