Refactor to store multiple reports per patchset
The current implementation only saves one test report per patchset
however each patchset could contain multiple reports. The use case for
that is when a job fails, not due to a code or test error, but instead
due to some flakyness with the CI system that ran the job.  In that
case you may want to re-run the job(s) on the same patchset and save
that as a seperate report.

Change-Id: I468b1c67afa8e0c6e776bc6c221f14f8397f2c89