Add a screen to display all jobs for a patchset
There can be multiple reports per patchset however the change screen
only shows a single report (the most current one).  We add a link
to allow users to view all reports for a patchet in a seperate screen.
We use a seperate screen because the list of jobs from multiple
reports can be very long and there's not enough space on the change
screen to show them all.

Change-Id: Idb74d1b19095dce8452b898cd0c9708534173e71