Add a job summary table
Add a job summary table to the change screen showing the number of
passed, failed and unstable jobs.  This summary info is only
for the most current report.

This change also adds a verify-status~stats rest endpoint to
facilitate the visualization of the jobs summary.  It also might
be nice to include other jobs stats in the future.

Change-Id: Ieb86101217c23c91e3f5cb208577031611bc6295