Set up test infrastructure
Database is populated in init plugin step. There is no way to use
init step code in plugin test class, because of deep Guice graph.
Instead we create the table manually. This has a drawback that when
the entity class is modified, the test must be adjusted as well.
We don't expect this to happen very frequently, though.

To run the test:

  $ buck test --include verify-status
  [-] TESTING...FINISHED 11.6s (1 PASS/0 FAIL)
  RESULTS FOR ALL TESTS
  PASS     10.6s  1 Passed   0 Skipped   0 Failed
  com.googlesource.gerrit.plugins.verifystatus.VerifyStatusIT
  TESTS PASSED

Change-Id: I36f59810d5a3e74bdc163ba0790cda9d9786ae7e