Add config option for panels jobs ordering
Right now jobs ordering for JobsPanel and JobsDropDownPanel is
hardcoded as 'REPORTER' which is inconvenient in case when you
have some CI pipelines in place and so expect status to be
sorted by report time to be in accordance with pipeline flow.
So changes here introduce two more config options to control jobs
ordering in plugin panels. Possible sort options are REPORTER, NAME
and DATE. Default is REPORTER so the change is fully backward
compatible. Config example:
  [plugin "@PLUGIN@"]
     sortJobsDropDownPanel = NAME
     sortJobsPane = DATE

Change-Id: I27605c35b731bb088346aabe056c67eef4905f54
Signed-off-by: Sergii Kipot <kipot.sergey@gmail.com>