Add config option to control jobs panel position
If the list of jobs reported by CI is big it eats a lot of vertical
space just below review score which is not very nice.
This change introduces config option ShowJobsBelowRelatedInfoBlock
to control the position of jobs panel on the review page. If option is
set to 'true' the panel will be positioned just below related info block
(space on the right from review info). By default the option is set to
'false' which means current behavior (position below review scores).
Usage example:
  [plugin "verify-status"]
     ShowJobsBelowRelatedInfoBlock = true

Change-Id: Ia9a455388142982066f44b76116cba776fe0d149
Signed-off-by: Sergii Kipot <kipot.sergey@gmail.com>