Add configurable support for in-progress status
With long CI pipelines it's hard to detect the current pipeline
status with only passed,failed and unstable statuses. For instance if
jobs are re-triggered within the same patchset there is no easy way to
show this as we are still limited to only current three statuses but
they are not relevant if task is still in progress.
This change introduces config option enableInProgressStatus to enable
visualization of new 'in-progress' status. It has its own gif icon
shown in the same place as other statuses. If option is set to 'true'
in-progress status will be shown for all jobs reported with value=2.
By default the option is set to 'false' and nothing is changed then
comparing to current behavior.
Usage example:
  [plugin "verify-status"]
     enableInProgressStatus = true

Change-Id: I34cebc03c33ab0c55ff4079031b734d48f25765e
Signed-off-by: Sergii Kipot <kipot.sergey@gmail.com>