Update patch set 7
Patch Set 7:

(10 comments)

As this patch is more or less the bare minimum to get this plugin working on 2.16 the display is not full-featured. This would be another change, but I currently do not have the time to do this. But it's open-source, so if you need it you can supply a patch.

Patch-set: 7
Reviewer: Gerrit User 1011624 <1011624@173816e5-2b9a-37c3-8a2e-48639d4f1153>